modify: signin, signout 코드 정리
