+++ src/api/api.ts
@@ -1,19 +1,15 @@
 import { API_BASE_URL } from '../constants/api';
 import axios from 'axios';
-
 //OAuth 프레임워크에서 Bearer 사용. OAuth는 프레임워크 사용하는지 백엔드에 확인필요
 //TODO 저장된 토큰 가져오는 getToken() 함수 구현 해야함
-
-// const USER_TOKEN = `Bearer ${getToken('authorization')}`;
-const USER_TOKEN = 'Bearer authorization';
 
 export const apiInstance = axios.create({
   baseURL: `${API_BASE_URL}/`,
   timeout: 1000,
-  headers: { 'Content-Type': 'application/json', authorization: USER_TOKEN },
+  headers: { 'Content-Type': 'application/json' },
 
   // 서버와 클라이언트가 다른 도메인일 경우 필수
-  // withCredentials: true,
+  withCredentials: true,
 });
 
 apiInstance.interceptors.request.use(
